+++ src/bindEventListeners.js
@@ -9,26 +9,26 @@
  */
 export default function bindEventListeners($el, scope = $el, settings = {}) {
   const defaults = {
-    pattern: /^@(?<event>[^.]+).?(?<modifier>.+)?/,
+    pattern: /^@(?:(?<target>window|document)\.)?(?<event>[^.]+)(?:\.(?<modifier>.+))?/i,
     cleanUp: true,
   };
 
   const { pattern, cleanUp } = { ...defaults, ...settings };
   const walker = document.createTreeWalker($el, NodeFilter.SHOW_ELEMENT);
-  const get = (key, target) => key.split('.').reduce((o, i) => (o || {})[i], target);
+  const get = (target, key) => key.split('.').reduce((o, i) => o?.[i], target);
 
   do {
     const $currentNode = walker.currentNode;
     const matches = findAttributes($currentNode, pattern);
 
     matches.forEach(({ name, value }) => {
-      const { event, modifier = '' } = name.match(pattern).groups || {};
+      const { target = '', event, modifier = '' } = name.match(pattern).groups || {};
 
       if (event === undefined) {
         throw new Error('Pattern must include named group “event”.');
       }
 
-      const fn = get(value, scope);
+      const fn = get(scope, value);
 
       if (typeof fn !== 'function') {
         throw new Error(`Event listener function “${value}” is invalid or undefined for this element.`);
@@ -38,12 +38,14 @@ export default function bindEventListeners($el, scope = $el, settings = {}) {
         $currentNode.removeAttribute(name);
       }
 
-      const isWindow = modifier.includes('window');
-      const isDocument = modifier.includes('document');
+      const isWindow = target.toLowerCase() === 'window';
+      const isDocument = target.toLowerCase() === 'document';
       const isAway = modifier.includes('away');
       const isPrevent = modifier.includes('prevent');
       const isOnce = modifier.includes('once');
       const isSelf = modifier.includes('self');
+      const isCapture = modifier.includes('capture');
+      const isPassive = modifier.includes('passive');
 
       const $target = (isWindow && window) || (isDocument && document) || (isAway && window) || $currentNode;
 
@@ -63,7 +65,10 @@ export default function bindEventListeners($el, scope = $el, settings = {}) {
         }
       };
 
-      $target.addEventListener(event, handler);
+      $target.addEventListener(event, handler, {
+        capture: isCapture,
+        passive: isPassive,
+      });
     });
   } while (walker.nextNode() !== null);
 }